--- src/components/index.js
@@ -1,5 +1,0 @@
-export * from './InputGroup';
-export * from './Login';
-export * from './Signup';
-export * from './Todo';
-export * from './TodoList';
