--- src/pages/index.js
@@ -1,7 +1,0 @@
-// export { default as LoginPage } from './LoginPage';
-// export { default as SignupPage } from './SignupPage';
-// export { default as TodoPage } from './TodoPage';
-
-export * from './LoginPage.jsx';
-export * from './SignupPage.jsx';
-export * from './TodoPage.jsx';
